==> 06_filter_relevant_results.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
from gensim import corpora, models, similarities
import numpy as np
import ast

## We import the claim and the snippets

df_bow = pd.read_csv('data/02_bow_snippets_claims.csv')[['snippet_id','claim_id','snippet_content','claim_content']]
df_bow.columns = ['snippet_id','claim_id','snippet_bow','claim_bow']

## We evaluate strings from snippets and claims as list of terms

df_bow['snippet_bow'] = df_bow['snippet_bow'].apply(lambda st: ast.literal_eval(st))
df_bow['claim_bow'] = df_bow['claim_bow'].apply(lambda st: ast.literal_eval(st))

## We transform list of terms into list of indices using the chosen dictionary

dictionary = corpora.Dictionary.load('dictionaries/02_30.0.dict')

df_bow['snippet_bow'] = df_bow['snippet_bow'].apply(lambda x: dictionary.doc2bow(x))
df_bow['claim_bow'] = df_bow['claim_bow'].apply(lambda x: dictionary.doc2bow(x))

## We apply the chosen topic model on list of indices from snippets and claims 

lda = models.LdaModel.load('topic_models/03_30.0_200.lda')
df_bow['snippet_bow'] = df_bow['snippet_bow'].apply(lambda x: lda[x])
df_bow['claim_bow'] = df_bow['claim_bow'].apply(lambda x: lda[x])

## For each pair of claim/snippet, we compute the similarity measure

dic_similarities = {}
dic_similarities['snippet_id'] = list(df_bow[['claim_id','snippet_id']]\
.sort_values(by=['claim_id','snippet_id'],ascending=False)['snippet_id'].as_matrix())
dic_similarities['similarity_score']=[]

claim_ids = sorted(list(df_bow['claim_id'].unique()),reverse=True)

for claim_id in claim_ids:
    df_ = df_bow[df_merge['claim_id']==claim_id][['snippet_id','snippet_bow','claim_bow']]\
    .sort_values(by=['snippet_id'],ascending=False)
    claim_bow = list(df_['claim_bow'].as_matrix())[0]
    snippet_bow = list(df_['snippet_bow'].as_matrix())
    lda_index = similarities.MatrixSimilarity(snippet_bow)
    max_ = np.max([x[0] for lst in snippet_bow for x in lst])
    sims_lda = lda_index[list(filter(lambda x: x[0]<=max_,claim_bow))]
    for i in range(0,len(snippet_bow)):
        dic_similarities['similarity_score'].append(sims_lda[i])
df_similarities = pd.DataFrame(dic_similarities)  

df_merge = pd.merge(df_bow,df_similarities)

## We keep only snippets with a similarity score above 0.65 to ensure that the snippets are relevant

df_filter = df_merge[df_merge['similarity_score']>=0.65]

## We import data associated to the selected snippets

df_snippets_2 = pd.read_csv('data/00_dataset.csv')
df_final = pd.merge(df_snippets_2,df_filter[['snippet_id','similarity_score']])

## We export the data

df_final[['claim_id','snippet_content',
       'snippet_date', 'snippet_id', 'snippet_pagenum', 'snippet_title',
       'snippet_url', 'claim_content', 'claim_date', 'claim_label',
       'claim_tag', 'claim_url', 'similarity_score']].to_csv('data/06_relevent_snippets.csv', index=False)
