==> 07_build_dataset.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
from gensim import corpora
import numpy as np
import ast
import os
from bs4 import BeautifulSoup
import re
import urllib.parse

## We list all documents that we retrieved using http requests

documents = os.listdir('relevant_documents')

## How to extract texts from html documents

def visible(element):
    if element.parent.name in ['style', 'script', '[document]', 'head', 'title']:
        return False
    elif re.match('<!--.*-->', str(element.encode('utf-8'))):
        return False
    return True

def html_to_text(txt):
    soup = BeautifulSoup(txt,'html.parser')
    data = soup.findAll(text=True)
    res = list(filter(visible, data))
    return ' '.join(res)

#https://www.quora.com/How-can-I-extract-only-text-data-from-HTML-pages

## We extract texts from web documents and export them in a csv file

dic_documents = {'snippet_id':[], 'document_content':[]}

for file_name in documents:
    snippet_id = int(file_name.split('.')[0])
    file = open('relevant_documents/'+file_name)
    txt = html_to_text(file.read())
    file.close()
    dic_documents['snippet_id'].append(snippet_id)
    dic_documents['document_content'].append(txt)
    
df_documents = pd.DataFrame(dic_documents)

df_documents.to_csv('data/07_documents.csv', index=False)
