==> 11_features.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
from gensim import corpora, models, matutils
import numpy as np
import ast

## We import the data

df_doc = pd.read_csv('data/10_tfidf.csv')

## We evaluate strings as list of terms

for nb_word in sorted(list(np.arange(500,20000,500)),reverse=True):
    df_doc['document_bow_'+str(nb_word/1000)] = df_doc['document_bow_'+str(nb_word/1000)].apply(lambda st: ast.literal_eval(st))
    df_doc['document_tfidf_'+str(nb_word/1000)] = df_doc['document_tfidf_'+str(nb_word/1000)].apply(lambda st: ast.literal_eval(st))
for nb_word in sorted(list(np.arange(60,186,10)),reverse=True):    
    df_doc['claim_bow_'+str(nb_word/1000)] = df_doc['claim_bow_'+str(nb_word/1000)].apply(lambda st: ast.literal_eval(st))
    df_doc['claim_tfidf_'+str(nb_word/1000)] = df_doc['claim_tfidf_'+str(nb_word/1000)].apply(lambda st: ast.literal_eval(st))

## We apply each base and topic model on claims and documents

for nb_word in sorted(list(np.arange(500,20000,500)),reverse=True):
    for nb_topic in np.arange(20,310,20):
        lsi = models.LsiModel.load('topic_models/10_doc_'+str(nb_word/1000)+'_'+str(nb_topic)+'.lsi')
        rp = models.RpModel.load('topic_models/10_doc_'+str(nb_word/1000)+'_'+str(nb_topic)+'.rp')
        lda = models.RpModel.load('topic_models/10_doc_'+str(nb_word/1000)+'_'+str(nb_topic)+'.lda')
        
        df_doc['document_lsi_'+str(nb_word/1000)+'_'+str(nb_topic)] = \
        df_doc['document_tfidf_'+str(nb_word/1000)].apply(lambda x: lsi[x])
        df_doc['document_rp_'+str(nb_word/1000)+'_'+str(nb_topic)] = \
        df_doc['document_bow_'+str(nb_word/1000)].apply(lambda x: rp[x])
        df_doc['document_lda_'+str(nb_word/1000)+'_'+str(nb_topic)] = \
        df_doc['document_bow_'+str(nb_word/1000)].apply(lambda x: lda[x])

for nb_word in sorted(list(np.arange(60,186,10)),reverse=True):
    for nb_topic in np.arange(10,60,10):
        lsi = models.LsiModel.load('topic_models/10_claim_'+str(nb_word/1000)+'_'+str(nb_topic)+'.lsi')
        rp = models.RpModel.load('topic_models/10_claim_'+str(nb_word/1000)+'_'+str(nb_topic)+'.rp')
        lda = models.RpModel.load('topic_models/10_claim_'+str(nb_word/1000)+'_'+str(nb_topic)+'.lda')
        
        df_doc['claim_lsi_'+str(nb_word/1000)+'_'+str(nb_topic)] = \
        df_doc['claim_tfidf_'+str(nb_word/1000)].apply(lambda x: lsi[x])
        df_doc['claim_rp_'+str(nb_word/1000)+'_'+str(nb_topic)] = \
        df_doc['claim_bow_'+str(nb_word/1000)].apply(lambda x: rp[x])
        df_doc['claim_lda_'+str(nb_word/1000)+'_'+str(nb_topic)] = \
        df_doc['claim_bow_'+str(nb_word/1000)].apply(lambda x: lda[x])

## We can see base and topic model representations as features
## We create a list containing the names of base features
## and another containing the names of topic features

feature_categories_document = []
for nb_word in sorted(list(np.arange(500,20000,500)),reverse=True):
    for f in ['bow','tfidf']:
        feature_categories_document.append('document_'+f+'_'+str(nb_word/1000))
    for nb_topic in np.arange(20,310,20):
        for f in ['lsi','lda','rp']:
            feature_categories_document.append('document_'+f+'_'+str(nb_word/1000)+'_'+str(nb_topic))
feature_categories_claim = []            
for nb_word in sorted(list(np.arange(60,186,10)),reverse=True):
    for f in ['bow','tfidf']:
        feature_categories_claim.append('claim_'+f+'_'+str(nb_word/1000))
    for nb_topic in np.arange(10,60,10):
        for f in ['lsi','lda','rp']:
            feature_categories_claim.append('claim_'+f+'_'+str(nb_word/1000)+'_'+str(nb_topic))

## We select the features with the associated labels that we export

df_doc[['claim_id','claim_label']+feature_categories_document+feature_categories_claim].to_csv('data/11_features.csv')
